requirements.txt
Django==1.11.7
django-filter==1.1.0
django-treebeard==4.1.2
djangorestframework==3.7.3
jsonfield==2.0.2
mysqlclient==1.3.12
pytz==2017.3
treebeard==0.1.1
